Finalizing switch to the tview
