+++ utils/cli.go
@@ -3,7 +3,19 @@ package utils
 import (
 	"os"
 	"os/exec"
+	"strings"
 )
+
+// CleanArgs trims the whitespace from the given set of strings.
+func CleanArgs(s []string) []string {
+	var r []string
+	for _, str := range s {
+		if str != "" {
+			r = append(r, strings.Trim(str, " "))
+		}
+	}
+	return r
+}
 
 // runCmd runs a given shell command in the current tty
 func runCmd(cmdStr string, args ...string) error {
